+++ JASP-Desktop/widgets/itemmodelselectitem.cpp
@@ -40,7 +40,7 @@ void ItemModelSelectItem::bindTo(Option *option)
 
 	vector<string> options = _boundTo->options();
 
-	for (int i = 0; i < options.size(); i++)
+	for (size_t i = 0; i < options.size(); i++)
 	{
 		if (options.at(i) == _boundTo->value())
 		{
